geo: move TestClipRectByEWKT test to geoindex
Trying to make the geo package not depend on GEOS, and this is one of
the required steps.

Release note: None
