+++ pkg/geo/geoindex/testdata/clip
